// web/src/api/generated/models/deliveryResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 1.0.0
 */
import type { DeliveryResultHeaders } from './deliveryResultHeaders';
import type { DeliveryResultPayload } from './deliveryResultPayload';
import type { DeliveryResultResponseHeaders } from './deliveryResultResponseHeaders';

export interface DeliveryResult {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  attemptId: string;
  broadcasterUserId: number;
  destination: string;
  durationMs: number;
  error?: string;
  eventType: string;
  eventVersion: number;
  generatedEventId: string;
  headers: DeliveryResultHeaders;
  logicalSubscription: string;
  messageId: string;
  messageTimestamp: string;
  method: string;
  outcome: string;
  payload: DeliveryResultPayload;
  rawBody: string;
  rawHttp: string;
  replayMode?: string;
  replayOfId?: string;
  responseBody?: string;
  responseHeaders?: DeliveryResultResponseHeaders;
  responseTruncated: boolean;
  runId: string;
  status?: number;
  subscriptionId: string;
  transportStartedAt: string;
  url: string;
}