// web/src/api/generated/models/eventDeliveryRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kick Sim Studio API
 * OpenAPI spec version: 1.0.0
 */
import type { EventDeliveryRequestPayload } from './eventDeliveryRequestPayload';

export interface EventDeliveryRequest {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  destination?: string;
  destinationUrl?: string;
  /** @minLength 1 */
  eventType: string;
  /** @minimum 1 */
  eventVersion: number;
  payload: EventDeliveryRequestPayload;
  subscriptionId?: string;
}